get html in gridLayout page
